requirements.txt
Flask
flask-limiter
flask-cors
python-dotenv
requests
openai
